code now working!
found out what the error was! there was  two more declerations for the name of the color sensor and that was the error - chaneged the values of the ball drop autonomus to 100 blue and 100 red - true value but not tested.
can drive! wheel rotation is working but i forgot to test if it was forward or backwards( lol). but the code works and now we can go back to work!